Add link to title and finish updating filepaths and renaming files
