+++ src/containers/Forms/SignUpForm/SignUpForm.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import { addUser } from '../../util/apiCalls';
-import { validUser } from '../../actions/index';
+import { addUser } from '../../../util/apiCalls';
+import { validUser } from '../../../actions/index';
 import { connect } from 'react-redux';
 import { Redirect} from 'react-router-dom'
 import './SignUpForm.css'
